Seperate project into API and Launcher
