+++ api/src/main/java/com/mineteria/ignite/api/Platform.java
@@ -1,0 +1,4 @@
+package com.mineteria.ignite.api;
+
+public interface Platform {
+}
