+++ api/src/main/java/com/mineteria/ignite/api/mod/ModContainer.java
@@ -22,38 +22,55 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.mineteria.ignite;
+package com.mineteria.ignite.api.mod;
 
-import com.google.gson.Gson;
-import com.mineteria.ignite.mod.ModEngine;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.checkerframework.checker.nullness.qual.NonNull;
-import org.spongepowered.asm.launch.MixinBootstrap;
+import org.checkerframework.checker.nullness.qual.Nullable;
 
-public final class IgniteCore {
-  public static final IgniteCore INSTANCE = new IgniteCore();
+import java.util.Objects;
 
-  private final Logger logger = LogManager.getLogger("IgniteCore");
-  private final Gson gson = new Gson();
+public final class ModContainer {
+  private final ModResource resource;
+  private final ModConfig config;
 
-  private final ModEngine modEngine;
-
-  /* package */ IgniteCore() {
-    this.modEngine = new ModEngine(this);
-
-    MixinBootstrap.init();
+  public ModContainer(final @NonNull ModResource resource,
+                      final @NonNull ModConfig config) {
+    this.resource = resource;
+    this.config = config;
   }
 
-  public @NonNull Logger getLogger() {
-    return this.logger;
+  public @NonNull String getId() {
+    return this.config.getId();
   }
 
-  public @NonNull Gson getGson() {
-    return this.gson;
+  public @NonNull String getVersion() {
+    return this.config.getVersion();
   }
 
-  public @NonNull ModEngine getEngine() {
-    return this.modEngine;
+  public @NonNull ModResource getResource() {
+    return this.resource;
+  }
+
+  public @NonNull ModConfig getConfig() {
+    return this.config;
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(this.resource, this.config);
+  }
+
+  @Override
+  public boolean equals(final @Nullable Object other) {
+    if (this == other) return true;
+    if (!(other instanceof ModContainer)) return false;
+    final ModContainer that = (ModContainer) other;
+    return Objects.equals(this.resource, that.resource)
+      && Objects.equals(this.config, that.config);
+  }
+
+  @Override
+  public @NonNull String toString() {
+    return this.getId() + "@" + this.getVersion();
   }
 }
