+++ launcher/src/main/java/com/mineteria/ignite/mixin/IgniteMixinPlatformService.java
@@ -24,34 +24,32 @@
  */
 package com.mineteria.ignite.mixin;
 
-import com.google.common.collect.ImmutableList;
 import org.checkerframework.checker.nullness.qual.NonNull;
-import org.spongepowered.asm.launch.platform.container.ContainerHandleModLauncher;
-import org.spongepowered.asm.service.modlauncher.MixinServiceModLauncher;
+import org.checkerframework.checker.nullness.qual.Nullable;
+import org.spongepowered.asm.launch.platform.IMixinPlatformServiceAgent;
+import org.spongepowered.asm.launch.platform.MixinPlatformAgentAbstract;
+import org.spongepowered.asm.launch.platform.MixinPlatformManager;
+import org.spongepowered.asm.launch.platform.container.IContainerHandle;
 
 import java.util.Collection;
 
-public final class IgniteMixinService extends MixinServiceModLauncher {
+@SuppressWarnings("unused")
+public final class IgniteMixinPlatformService extends MixinPlatformAgentAbstract implements IMixinPlatformServiceAgent {
   @Override
-  public boolean isValid() {
-    return true;
+  public void init() {}
+
+  @Override
+  public @NonNull AcceptResult accept(final @NonNull MixinPlatformManager manager, final @NonNull IContainerHandle handle) {
+    return AcceptResult.REJECTED;
   }
 
   @Override
-  public ContainerHandleModLauncher getPrimaryContainer() {
-    return new LauncherContainer(this.getName());
+  public @NonNull String getSideName() {
+    return "SERVER";
   }
 
   @Override
-  public Collection<String> getPlatformAgents() {
-    return ImmutableList.<String>of(
-      "com.mineteria.ignite.mixin.IgniteMixinPlatformService"
-    );
-  }
-
-  private static final class LauncherContainer extends ContainerHandleModLauncher {
-    public LauncherContainer(final @NonNull String name) {
-      super(name);
-    }
+  public @Nullable Collection<IContainerHandle> getMixinContainers() {
+    return null;
   }
 }
