+++ launcher/src/main/java/com/mineteria/ignite/util/IgniteConstants.java
