+++ launcher/src/main/resources/META-INF/services/cpw.mods.modlauncher.api.ILaunchHandlerService
