+++ launcher/src/main/resources/META-INF/services/org.spongepowered.asm.service.IMixinService
