+++ settings.gradle
@@ -1,3 +1,5 @@
 rootProject.name = "Ignite"
 
+include("api")
+include("launcher")
 include("example")
